+++ CHANGES.md
@@ -21,7 +21,6 @@ Roadmap
 * Enhancement of flash CLI utility, specify a max_gap/min_gap parameter.
 * Rewrite combination CLI to have a more simple usage.
 * Write a transposition CLI tool that rewrite data lines to columns.
-* Write a complement CLI tool that computes complement combinations of input file.
 * adding a minimal_cover CLI tool.
 * Move combination CLI to an ESM module.
 * Move filter CLI to an ESM module.
@@ -31,6 +30,7 @@ Roadmap
 
 2023-xx-xx, Version 0.0.1-alpha.10
 ==================================
+* Adding complement CLI utility.
 * Documentation update.